Migrate ServiceListenerTest
